Removed tangential link
Nothing about adding a group repo is currently described in the link provided.  I checked NPM (http://books.sonatype.com/nexus-book/reference/npm-grouping-registries.html) and they do not have that link so I am removing assuming that it's an oversight rather than a WIP.
Interestingly, I am then unsure if we have anywhere that describes creating a group (though groups are described in the link I removed).  This might be a minor hole.