Suggested strike
The testguide specifically says you cannot assign privileges to users, that you have to assign the privilidges to a role which then you can assign to a user.  My test of this showed this to be true.  While just a wording difference, I think not having privilege lessens the chance of confusion and IMO is better than the alternative which would be to explain this verbosely.