NEXUS-9781-rubygems-groupplus-revamp
* Revamped the RubyGems group section from NX2 port after rework in https://issues.sonatype.org/browse/NEXUS-9781
* Also reviewed hosted, using gem repositories and pushing gems which seemed thus far untouched
* Minor typo fixes
* Changed link in introduction (it looked like it was including information on hosted as written)
* 114ized some sections, will take another pass after PR is merged (so not to cause [more] visual confusion)

NOTE: I intentionally did not include images because I did not see the value.  One place I was wishy washy, I left the comment in there, the rest I struck.  If people believe worth feel free to note on PR and I can or add in yourself (now or later).