Unticketed search improvements
Noticed one while testing and more on review, none of these have
tickets associated that I know of.

* Added PyPI search specifics
* Removed npm redundancy: it listed generics as specifics where there
was no different from the generics.  If this cull is -1ed for any
reason, note that it is also incorrect.  It lists 'group' but npm shows
'scope'.
* Made npm search description consistent with others: adding format
* Removed extra newline
* Hiding incorrect age and popularity information and updated
screenshot until restored (ref
https://issues.sonatype.org/browse/NEXUS-10770)

Bonus:
* Adjusted some screenshot (non-search) sizing to make pdf seeable
* Made a 2 line comment onto 1 line (still 114ized)