Removing (minimum)s (#62)
I believe our intent when putting (minimum) in the table header was that if we had, for example, 3.3 which didn't require a new 2.14.2 to upgrade then it'd be "minimum".  However, this isn't currently the case, so the headers seem to have a strong potential to cause confusion.
For now, I am striking but for the future, if we use the word minimum, I suggest we put it in the row and not the headers...or come up with a better strategy.