Striking untrue paragraph (#126)
User noted in http://sonatype.zendesk.com/agent/tickets/20108.  As far as I know this is not a bug (or story; this is not something we intend to do at this juncture...maybe at first it was).  Feel free to include in review if you feel otherwise and I can ticket.